refactoring: moved sessionSlice into session/model. Created public api for this slice.
