+++ src/entities/session/index.ts
@@ -1,1 +1,3 @@
 export * from './api/loginApi';
+export * from './model/types';
+export * from './model/sessionSlice';
