+++ src/shared/api/index.ts
@@ -1,8 +1,8 @@
 export * from './model/types';
-export * from './model/authSlice';
+export * from '@entities/session/model/sessionSlice';
 export * as ENDPOINTS from './config/endpoints';
 export * as METHODS from './config/methods';
 export * as TAGS from './config/invalidation-tags';
 
-export { default as authSlice } from './model/authSlice';
+export { default as authSlice } from '@entities/session/model/sessionSlice';
 export { default as baseAPI } from './baseApi';
