+++ src/shared/api/model/unauthorizedHandler.ts
@@ -1,11 +1,11 @@
 import type { FetchBaseQueryError } from '@reduxjs/toolkit/query';
 import { ROUTES } from '@shared/router';
-import { removeStoredToken } from './tokenStorageHandlers';
+import { removeTokenFromLocalStorage } from './tokenLocalStorageHandlers';
 
 const unauthorizedHandler = (error: FetchBaseQueryError | undefined): void => {
   if (error?.status === 401) {
     console.warn('🔄 Unauthorized! Removing token and redirecting to login.');
-    removeStoredToken();
+    removeTokenFromLocalStorage();
     window.location.replace(ROUTES.LOGIN);
   }
 };
